crossplat round 2c: bun_runtime win-msvc 352→328 — dir_iterator WrappedSelect cross-platform marker for NewWrappedIterator/iterate where-bound; node_fs wrap_to per-platform u32/u8, mkdir_os_path→sys::mkdir_w, openat_os_path→openat_windows, to_sys_time_like f64→{sec,nsec}, event_loop_mut for enqueue_task; sys_uv ftruncate i64; blob copy/write_file virtual_machine.unwrap().as_ref().event_loop_handle deref

+++ src/sys/sys_uv.rs
@@ -342,11 +342,14 @@ pub fn symlink_uv(target: &ZStr, new_path: &ZStr, flags: c_int) -> Result<()> {
     }
 }
 
-pub fn ftruncate(fd: Fd, size: isize) -> Result<()> {
+pub fn ftruncate(fd: Fd, size: i64) -> Result<()> {
+    // Zig spec types `size: isize` (= i64 on every supported Windows target);
+    // accept `i64` directly so this matches the cross-platform `bun_sys::ftruncate`
+    // signature and callers don't need a per-platform `as isize` cast.
     let uv_fd = fd.uv();
     let mut req = uv::fs_t::uninitialized();
     // SAFETY: synchronous libuv fs call; req lives on the stack for the duration.
-    let rc = unsafe { uv::uv_fs_ftruncate(uv::Loop::get(), &mut req, uv_fd, size as i64, None) };
+    let rc = unsafe { uv::uv_fs_ftruncate(uv::Loop::get(), &mut req, uv_fd, size, None) };
 
     log!("uv ftruncate({}, {}) = {}", uv_fd, size, rc.int());
     if let Some(errno) = rc.errno() {
